services/simcct-server/requirements.txt
coverage==4.5.3
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-Testing==0.7.1
gunicorn==19.9.0
pymongo==3.8.0
prettytable==0.7.2
python-dateutil==2.8.0
python-dotenv==0.10.3